Set default open file limit to something more sensible
A bug (?) in slapd cause it to use a whole bunch of memory if the nofile
ulimit is set too high.

Set it to the smaller of 16384 or the preexisting ulimit (as set by the
`--ulimit` option to `docker run`, for example).

If you want to set it higher you'll need to set the SLAPD_NOFILE_SOFT
environment variable.

Closes #8.